Write tests for GeneralTripViewModel, fix updating trip list
