+++ app/build.gradle
@@ -70,6 +70,7 @@ dependencies {
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
     testImplementation 'io.mockk:mockk:1.12.3'
     testImplementation 'org.jetbrains.kotlinx:kotlinx-coroutines-test:1.5.0'
+    testImplementation 'app.cash.turbine:turbine:0.7.0'
 
     /** firebase **/
     implementation platform('com.google.firebase:firebase-bom:29.0.4')
