+++ app/src/test/java/com/example/hikeculator/presentation/general_trip_list/GeneralTripViewModelTest.kt
@@ -1,37 +1,29 @@
 package com.example.hikeculator.presentation.general_trip_list
 
-import android.util.Log
+import app.cash.turbine.test
+import com.example.hikeculator.MainCoroutineRule
 import com.example.hikeculator.domain.entities.Trip
-import com.example.hikeculator.domain.entities.User
-import com.example.hikeculator.domain.enums.TripDifficultyCategory
-import com.example.hikeculator.domain.enums.TripSeason
-import com.example.hikeculator.domain.enums.TripType
 import com.example.hikeculator.domain.interactors.ProvisionBagInteractor
 import com.example.hikeculator.domain.interactors.TripDayInteractor
 import com.example.hikeculator.domain.interactors.TripInteractor
 import com.example.hikeculator.domain.interactors.UserProfileInteractor
-import io.mockk.*
-import kotlinx.coroutines.*
-import kotlinx.coroutines.flow.Flow
+import io.mockk.coEvery
+import io.mockk.coVerifyAll
+import io.mockk.every
+import io.mockk.mockk
+import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.flow
-import kotlinx.coroutines.flow.flowOf
-import kotlinx.coroutines.test.*
-import org.junit.*
-import org.junit.Assert.*
+import kotlinx.coroutines.runBlocking
+import org.junit.Assert.assertNotNull
+import org.junit.Before
+import org.junit.Rule
+import org.junit.Test
 
-import org.junit.rules.TestWatcher
-import org.junit.runner.Description
-
-//@get:Rule
-//val coroutineRule = GeneralTripViewModelTest.MainCoroutineRule()
 @ExperimentalCoroutinesApi
 class GeneralTripViewModelTest {
 
     @get:Rule
     var coroutineRule = MainCoroutineRule()
-
-    @ExperimentalCoroutinesApi
-    val dispatcher = TestCoroutineDispatcher()
 
     private val tripInteractor: TripInteractor = mockk(relaxUnitFun = true)
     private val tripDayInteractor: TripDayInteractor = mockk(relaxUnitFun = true)
@@ -45,7 +37,6 @@ class GeneralTripViewModelTest {
     @ExperimentalCoroutinesApi
     @Before
     fun setUp() {
-        Dispatchers.setMain(dispatcher = dispatcher)
         viewModel = GeneralTripViewModel(
             tripInteractor = tripInteractor,
             tripDayInteractor = tripDayInteractor,
@@ -54,24 +45,13 @@ class GeneralTripViewModelTest {
         )
     }
 
-    @ExperimentalCoroutinesApi
-    @After
-    fun tearDown() {
-        Dispatchers.resetMain()
-        dispatcher.cleanupTestCoroutines()
-//        dispatcher.close()
-    }
-
     @Test
-    fun `invoke deleting day collection and provision bag and trip on deleting trip`() =
+    fun `invoke deleting day collection, provision bag and trip on deleting trip`() =
         runBlocking {
-
             val tripId = "some id"
-
             val trip = mockk<Trip> { every { id } returns tripId }
 
             viewModel.deleteTrip(trip = trip)
-
 
             coVerifyAll() {
                 tripDayInteractor.removeTripDayCollection(tripId = tripId)
@@ -80,21 +60,19 @@ class GeneralTripViewModelTest {
             }
         }
 
-}
+    @Test
+    fun `getting message when deleting trip was failed`() {
+        runBlocking {
+            val trip = mockk<Trip>()
 
-@ExperimentalCoroutinesApi
-class MainCoroutineRule(
-    private val dispatcher: TestCoroutineDispatcher = TestCoroutineDispatcher()
-) : TestWatcher(), TestCoroutineScope by TestCoroutineScope(dispatcher) {
+            coEvery { tripDayInteractor.removeTripDayCollection(any()) } throws Exception()
+            coEvery { provisionBagInteractor.removeProvisionBag(tripId = any()) } throws Exception()
+            coEvery { tripDayInteractor.removeTripDayCollection(tripId = any()) } throws Exception()
 
-    override fun starting(description: Description?) {
-        super.starting(description)
-        Dispatchers.setMain(dispatcher)
-    }
-
-    override fun finished(description: Description?) {
-        super.finished(description)
-        cleanupTestCoroutines()
-        Dispatchers.resetMain()
+            viewModel.problemMessage.test {
+                viewModel.deleteTrip(trip = trip)
+                assertNotNull(awaitItem())
+            }
+        }
     }
 }
